ci: remove activate_credentials_file arg from google cloud auth (#1983)
CI[ was raising a warning
](https://github.com/Unstructured-IO/unstructured/actions/runs/6725942303)in
the workflow annotations (scroll to the bottom).

It looks like this is not a supported argument for the version that we
are targeting. Since it looks like this is probably just getting ignored
anyways, removing.

## Testing
Note the [CI workflow
run](https://github.com/Unstructured-IO/unstructured/actions/runs/6726986305)
from this PR succeeded and no longer has those annotations.