fix(html): invalid <br> elements in text_as_html
In an apparent attempt to replace newlines ("\n") in cell text with a
`<br/>` element (a reasonable idea), all newlines in `.text_as_html` are
replaced with `<br>`. The HTML generated by `tabulate` includes a
newline after each major element, so this inserts unwanted `<br>`
elements in invalide positions (like between `<tbody>` and `<tr>`).

Remove this for now and replace the desired part with our own HTML
generator in a following commit.