fix(html): text_as_html has whitespace padding
The current implementation uses `tabulate` to form the `.text_as_html`
metadata for an `HTMLTable` element. Because it is build for interactive
use, `tabulate` inserts padding such that columns align visually.

This whitespace is undesireable for downstream purposes where
compactness is most important.

Add a simple HTML table generator that conforms to our needs and use it
instead of `tabulate`.