chore: update api key link (#2182)
Update to API key link per Slack convo.

There may be more that's needed, @ron-unstructured please run with this
PR if it's the right solution, close it if not, or make changes if
needed. I haven't looked for other places where the link might need to
be changed. The link appears to work, but any further investigation or
testing is appreciated.

Co-authored-by: Ronny H <138828701+ron-unstructured@users.noreply.github.com>