rfctr: rename section -> pre-chunk throughout
The original naming for the pre-cursor to a chunk in `chunk_by_title()`
was conflated with the idea of how these element subsequences were
bounded (by document-section). I mistakenly picked that up as a
universal concept but in fact no notion of section arises in the
`by_character` chunking strategy.

Fix this misconception by using the name `pre-chunk` for this concept.